revno: 8722 message:   Small change ui of program function in mantenance-patient.
committer: Tran Chau <tran.hispvietnam@gmail.com>
branch nick: dhis2
timestamp: Fri 2012-10-26 10:26:57 +0700

+++ dhis-2/dhis-web/dhis-web-maintenance/dhis-web-maintenance-patient/src/main/java/org/hisp/dhis/patient/action/caseaggregation/TestCaseAggregationConditionAction.java
@@ -87,7 +87,7 @@ public class TestCaseAggregationConditionAction
         period.setStartDate( new Date() );
         period.setEndDate( new Date() );
 
-        Double value = aggregationConditionService.parseConditition( aggCondition, orgunit, period );
+        Integer value = aggregationConditionService.parseConditition( aggCondition, orgunit, period );
 
         if ( value == null )
         {
